Replace sort.Search with something faster.
* switched from int64 to uint (that's still 32-bits and gives us year range of 1970-2106)
* add yearSearch, that will interpolate the first guess and linear search the rest (seems slightly faster than binary search for the year case)
* implement our own binarySearch for uint and replace sort.Search (the biggest speed up by far)
* add Lookup tests

+++ fastcal.go
@@ -1,11 +1,9 @@
 package fastcal
 
 type Fastcal interface {
-  Set(int64)
-  Hour() int
-  Day() int
-  Month() int
-  Year() int
+	Set(uint)
+	Hour() int
+	Day() int
+	Month() int
+	Year() int
 }
-
-
